♻️ Refactor EnemyTypeC into separate module
Extracts EnemyTypeC class from GameCanvas component into dedicated entity file for better code organization and maintainability.

Removes 265 lines of inline class definition while preserving all functionality through proper import structure.
