+++ src/components/GameCanvas.tsx
@@ -5,6 +5,7 @@ import GameUI from "./GameUI";
 import Enemy from "../game/entities/Enemy";
 import EnemyTypeA from "../game/entities/EnemyTypeA";
 import EnemyTypeB from "../game/entities/EnemyTypeB";
+import EnemyTypeC from "../game/entities/EnemyTypeC";
 import Player from "../game/entities/Player";
 import SpawnSystem from "../game/systems/SpawnSystem";
 import WaveManager from "../game/systems/WaveManager";
@@ -284,265 +285,10 @@ const GameCanvas: React.FC<GameCanvasProps> = ({ backgroundSpeed = .75, onNaviga
     // e agora herda corretamente de AbstractEntity com todas as funcionalidades otimizadas.
 
     // =============================================
-    // ENEMY TYPE C
+    // ENEMY TYPE C - AGORA EM ARQUIVO SEPARADO
     // =============================================
-    class EnemyTypeC extends Enemy {
-      // Estados do inimigo
-      private state: 'entering' | 'moving_to_position' | 'shooting' | 'moving' | 'leaving' = 'entering';
-      
-      // Sistema de disparo
-      private fireTimer = 0;
-      private fireRate = 800; // 800ms entre tiros
-      private currentShots = 0;
-      private shotsPerBurstC = 5; // 5 tiros por rajada
-      
-      // Sistema de movimento
-      protected moveSpeed = 180; // Aumentado de 120 para 180
-      protected targetX = 0; // Alterado para protected
-      protected targetY = 0; // Alterado para protected
-      protected movementCountC = 0; // Alterado para protected
-      protected maxMovementsPerCycle = 3; // 3 movimentos por ciclo
-      private cycleCount = 0;
-      private maxCycles = 3; // 3 ciclos antes de sair
-      
-      constructor(scene: Phaser.Scene, x: number, y: number) {
-        super(scene, x, y, 'enemy-C-frame-1');
-        
-        // Stats específicos do Enemy Type C
-        this.hp = 1;
-        this.maxHp = 1;
-        this.speed = 120;
-        this.maxSpeed = 200;
-        this.shotsPerBurst = 5; // 5 tiros por rajada
-        this.maxShotsPerBurst = 8;
-        this.totalShots = 15; // Total de 15 tiros (3 ciclos * 5 tiros)
-        this.maxTotalShots = 25;
-        this.movementsBeforeLeaving = 3;
-        this.maxMovements = 5;
-        this.points = 500; // Inimigo tipo C dá 500 pontos
-        
-        // Configurar sprite
-        this.sprite.setScale(1);
-        this.sprite.setData('enemyType', 'C');
-        this.sprite.setData('hp', this.hp);
-        
-        // Definir primeira posição aleatória
-        this.setNextPosition();
-      }
-      
-      private setNextPosition() {
-        // Gerar posição aleatória inicial na área de jogo
-        const preferredX = Phaser.Math.Between(150, 650);
-        const preferredY = Phaser.Math.Between(120, 250);
-        
-        // Encontrar posição livre
-        const gameScene = this.scene as GameScene;
-        const freePosition = gameScene.findFreePosition(preferredX, preferredY);
-        
-        this.targetX = freePosition.x;
-        this.targetY = freePosition.y;
-      }
-      
-      private setExitTarget() {
-        // Sair pela parte superior da tela
-        this.targetX = this.sprite.x;
-        this.targetY = -100; // Sair pelo topo
-      }
-      
-      move() {
-        const deltaTime = this.scene.game.loop.delta / 1000;
-        
-        // Calcular direção para o alvo
-        const directionX = this.targetX - this.sprite.x;
-        const directionY = this.targetY - this.sprite.y;
-        const distance = Math.sqrt(directionX * directionX + directionY * directionY);
-        
-        // Se chegou perto o suficiente do alvo
-        if (distance < 10) {
-          this.onReachedTarget();
-          return;
-        }
-        
-        // Normalizar direção e aplicar velocidade
-        const normalizedX = directionX / distance;
-        const normalizedY = directionY / distance;
-        
-        this.sprite.x += normalizedX * this.moveSpeed * deltaTime;
-        this.sprite.y += normalizedY * this.moveSpeed * deltaTime;
-        
-        // Atualizar animação de movimento
-        this.updateMovementAnimation(normalizedX);
-      }
-      
-      private onReachedTarget() {
-        switch (this.state) {
-          case "moving_to_position": {
-            // Chegou na posição inicial, começar a atirar e reservar posição
-            this.state = "shooting";
-            this.resetShooting();
-
-            const gameScene = this.scene as GameScene;
-            gameScene.reservePosition(this.sprite.x, this.sprite.y, this.id);
-            break;
-          }
-
-          case "moving": {
-            // Chegou na nova posição após completar tiros, começar a atirar novamente
-            this.state = "shooting";
-            this.resetShooting();
-
-            const gameScene = this.scene as GameScene;
-            gameScene.reservePosition(this.sprite.x, this.sprite.y, this.id);
-            break;
-          }
-
-          case "leaving":
-            // Saiu da tela
-            this.isDestroyed = true;
-            // Não marcar isKilledByPlayer como true - inimigo saiu da tela
-            break;
-        }
-      }
-      
-      shoot() {
-        // Reproduzir som do tiro do inimigo
-        this.scene.sound.play('enemy-shoot', { volume: 0.2 });
-        
-        // Obter referência ao grupo de projéteis inimigos da scene
-        const gameScene = this.scene as GameScene;
-        if (!gameScene.enemyBullets) return;
-        
-        // Capturar posição atual do jogador no momento do disparo
-        if (!gameScene.player) return;
-
-        const playerX = gameScene.player.x;
-        const playerY = gameScene.player.y;
-        
-        // Calcular direção para o jogador
-        const directionX = playerX - this.sprite.x;
-        const directionY = playerY - this.sprite.y;
-        const distance = Math.sqrt(directionX * directionX + directionY * directionY);
-        
-        // Normalizar direção
-        const normalizedX = directionX / distance;
-        const normalizedY = directionY / distance;
-        
-        // Tentar criar projétil
-        let bullet = gameScene.enemyBullets.get(this.sprite.x, this.sprite.y + 20, 'enemy-fire');
-        
-        // Sistema de fallback para garantir que o projétil seja criado
-        if (!bullet) {
-          // Limpar projéteis que saíram da tela para liberar espaço no grupo
-          gameScene.enemyBullets.children.entries.forEach((child: Phaser.GameObjects.GameObject) => {
-            const sprite = child as Phaser.GameObjects.Sprite;
-            if (sprite.active && (sprite.y > 650 || sprite.y < -50 || sprite.x < -50 || sprite.x > 850)) {
-              sprite.setActive(false).setVisible(false);
-            }
-          });
-          
-          // Tentar novamente após limpeza
-          bullet = gameScene.enemyBullets.get(this.sprite.x, this.sprite.y + 20, 'enemy-fire');
-        }
-        
-        if (bullet) {
-          bullet.setActive(true);
-          bullet.setVisible(true);
-          bullet.setScale(0.8);
-          
-          // Velocidade do projétil direcionado
-          const speed = 280;
-          const velocityX = normalizedX * speed;
-          const velocityY = normalizedY * speed;
-          
-          bullet.setData('velocityX', velocityX);
-          bullet.setData('velocityY', velocityY);
-          bullet.setData('damage', 1);
-          bullet.setData('isEnemyBullet', true);
-          bullet.setData('isDirected', true); // Marcar como projétil direcionado
-        } else {
-          console.warn('Enemy C não conseguiu criar projétil direcionado');
-        }
-        
-        this.currentShots++;
-        this.shotsFired++;
-      }
-      
-      private updateMovementAnimation(directionX: number) {
-        // Atualizar frame baseado na direção do movimento
-        if (Math.abs(directionX) > 0.1) {
-          // Está se movendo horizontalmente - alternar entre frames 2 e 3
-          const frame = (this.scene.time.now % 600 < 300) ? 2 : 3;
-          this.sprite.setTexture(`enemy-C-frame-${frame}`);
-          
-          // Espelhar sprite baseado na direção
-          this.sprite.setFlipX(directionX > 0);
-        } else {
-          // Movimento apenas vertical ou parado - usar frame 1
-          this.sprite.setTexture('enemy-C-frame-1');
-          this.sprite.setFlipX(false);
-        }
-      }
-      
-      private resetShooting() {
-        this.currentShots = 0;
-        this.fireTimer = this.scene.time.now;
-        this.sprite.setTexture('enemy-C-frame-1'); // Frame de repouso
-      }
-      
-      update() {
-        super.update();
-        if (this.isDestroyed) return;
-        
-        const currentTime = this.scene.time.now;
-        
-        // Máquina de estados
-        switch (this.state) {
-          case 'entering':
-            // Verificar se entrou na tela
-            if (this.sprite.y > 30) {
-              this.state = 'moving_to_position';
-            } else {
-              // Continuar entrando (movimento para baixo)
-              this.sprite.y += this.moveSpeed * (this.scene.game.loop.delta / 1000);
-            }
-            break;
-            
-          case 'moving_to_position':
-          case 'moving':
-          case 'leaving':
-            this.move();
-            break;
-            
-          case 'shooting':
-            // Executar rajada de tiros
-            if (this.currentShots < this.shotsPerBurstC) {
-              if (currentTime - this.fireTimer > this.fireRate) {
-                this.shoot();
-                this.fireTimer = currentTime;
-              }
-            } else {
-              // Completou a rajada
-              this.onCompletedShooting();
-            }
-            break;
-        }
-      }
-      
-      private onCompletedShooting() {
-        this.cycleCount++;
-        
-        if (this.cycleCount >= this.maxCycles) {
-          // Completou todos os ciclos, sair da tela
-          this.setExitTarget();
-          this.state = 'leaving';
-        } else {
-          // Começar próximo ciclo de movimento
-          this.setNextPosition();
-          this.state = 'moving';
-          this.movementCountC = 0;
-        }
-      }
-    }
+    // A classe EnemyTypeC foi movida para src/game/entities/EnemyTypeC.ts
+    // e agora herda corretamente de AbstractEntity com todas as funcionalidades otimizadas.
 
     // =============================================
     // BOSS TYPE A
